requirements.txt
# Bot dependencies
py-clob-client
python-dotenv
requests
numpy

# API server
fastapi
uvicorn
jinja2

# Local dashboard
httpx
paramiko
